@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { View, Text } from 'react-native';
-import { Header } from "../common/index";
+import { Header } from "../../common/index";
 import LoginForm from "./LoginForm";
 
 export default class LoginPage extends Component {
